chore: complete infrastructure adapter consolidation
- Renamed password_hasher_service.py to password_hasher_adapter.py for consistency
- Added token_generator_adapter.py with proper adapter pattern implementation
- Updated token_generator_service.py with adapter pattern improvements
- Maintains backwards compatibility while improving architecture

These changes complete the infrastructure layer adapter consolidation ensuring
all external service adapters follow consistent naming and implementation patterns.
